Fix some issues with error highlighting:
 * MemoryTool was not recording errors passed to Tool.error() since they never made it to the error manager
 * In MemoryTool, split out 119 left-recursion errors to apply one error for each rule affected, and ensure sane document offsets for them, and modify error highlighting not to synthesize highlights for each one anymore
 * More logging and duplicate checking in error highlighting
 * Fix an indirect NPE if error or warning highlights were not yet computed in error highlighting
 * Ensure all errors and warnings from MemoryTool get highlightings applied for them
 * Prepare to make some of the inlining rule refactoring stuff public, so that we can try to add a left-recursion solver
 * More efficient collection of last modified info for AntlrGenerationResult
 * Ensure AntlrGenerationResult.cleanOldOutput() actually finds everything
 * Ensure that regeneration code always uses jfsSupplier, not the previous JFS, so it works if the IDE has been idle for a long time and the original JFS was deleted
 * Go back to creating an IsolationClassLoader per EmbeddedParser for now - not sure this was the root of problems of the compiler not finding Antlr classes, but it may have been - probably its internals getting closed
 * Don't use off-heap storage for mapping JFS's for now - problem was likely attempts to write to a closed JFS, but until that's tested, heap is safer
 * Ensure deleted files don't detonate if a JFSFileModifications tries to hash them

+++ antlr-editing-plugins/antlr-in-memory-build/src/main/java/org/nemesis/antlr/memory/AntlrGenerationResult.java
@@ -19,6 +19,7 @@ import com.mastfrog.graph.IntGraph;
 import com.mastfrog.graph.ObjectGraph;
 import com.mastfrog.util.path.UnixPath;
 import static com.mastfrog.util.preconditions.Checks.notNull;
+import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.charset.Charset;
 import java.nio.file.Path;
@@ -128,28 +129,37 @@ public final class AntlrGenerationResult implements ProcessingResult {
         this.hints = hints;
     }
 
-    public AntlrGenerationResult cleanOldOutput() {
-        Set<JFSFileObject> set = new HashSet<>();
-        for (JFSCoordinates.Resolvable f : modifiedFiles.keySet()) {
-            JFSFileObject ob = f.resolveOriginal();
-            if (ob != null) {
-                if (ob.storageKind().isMasqueraded()) {
-                    continue;
+    public AntlrGenerationResult cleanOldOutput() throws IOException {
+        if (!jfs.isEmpty()) {
+            jfs.whileWriteLocked(() -> {
+                Set<JFSCoordinates.Resolvable> all = new HashSet<>(newlyGeneratedFiles.size() +
+                        modifiedFiles.size());
+                all.addAll(newlyGeneratedFiles);
+                all.addAll(modifiedFiles.keySet());
+                Set<JFSFileObject> set = new HashSet<>();
+                for (JFSCoordinates.Resolvable f : all) {
+                    JFSFileObject ob = f.resolveOriginal();
+                    if (ob != null) {
+                        if (ob.storageKind().isMasqueraded()) {
+                            continue;
+                        }
+                        set.add(ob);
+                    }
                 }
-                set.add(ob);
-            }
-        }
-        for (JFSCoordinates.Resolvable f : newlyGeneratedFiles) {
-            JFSFileObject ob = f.resolveOriginal();
-            if (ob != null) {
-                if (ob.storageKind().isMasqueraded()) {
-                    continue;
+                for (JFSCoordinates.Resolvable f : newlyGeneratedFiles) {
+                    JFSFileObject ob = f.resolveOriginal();
+                    if (ob != null) {
+                        if (ob.storageKind().isMasqueraded()) {
+                            continue;
+                        }
+                        set.add(ob);
+                    }
                 }
-                set.add(ob);
-            }
-        }
-        for (JFSFileObject fo : set) {
-            fo.delete();
+                for (JFSFileObject fo : set) {
+                    fo.delete();
+                }
+                return null;
+            });
         }
         return this;
     }
